requirements.txt
SQLAlchemy==1.4.31
redis==6.2.0
minio == 7.1.10
pika==1.3.2
paramiko==3.5.1
psycopg2-binary==2.9.3
python-dotenv~=1.1.0
celery==5.5.3
